(gh-19) Flip puppetdb platform test in acceptance suite
Older platforms (currently Debian10 and Ubuntu18.04) do not have
postgresql packages available, so while we have openvox-server and
openvoxdb packages built, testing openvox-server with openvoxdb
integration needs to be skipped on these platforms since we can't
configure openvoxdb/postgresql there.

I inverted the function to test this so that we don't silently skip
testing openvoxdb integration on newly added platforms. Instead we now
explicitly restrict a subset of platforms where we know we can't
configure due to lack of postgresql.